Refactor Developer Settings to be more modular
