+++ packages/frontend-2/components/settings/user/developer/Applications/RevealSecretDialog.vue
