+++ packages/frontend-2/lib/settings/composables/menu.ts
@@ -1,7 +1,7 @@
 import type { SettingsMenuItems } from '~/lib/settings/helpers/types'
 import SettingsUserProfile from '~/components/settings/user/Profile.vue'
 import SettingsUserNotifications from '~/components/settings/user/Notifications.vue'
-import SettingsUserDeveloper from '~/components/settings/user/Developer.vue'
+import SettingsUserDeveloper from '~/components/settings/user/developer/Developer.vue'
 import SettingsUserEmails from '~/components/settings/user/Emails.vue'
 import SettingsServerGeneral from '~/components/settings/server/General.vue'
 import SettingsServerProjects from '~/components/settings/server/Projects.vue'
